+++ doc/version_history.rst
@@ -15,7 +15,8 @@ v0.4.0
 * Moved code to determine what sensor is connected from ts.ess.controller to ts.ess.common.
 * Moved BaseDevice and MockDevice from ts.ess.controller to ts.ess.common.
 * Removed all obsolete schema related code since it also is in ts.ess.common.
-* Updated the documentation to reflect all changes.
+* Updated the documentation to reflect all sensor and device code changes.
+* Moved most of the command handler code and the socket server unit test from ts.ess.controller to ts.ess.common.
 
 Requires: